[v0.91][maintainability] Split runtime_v2 standing module by responsibility (#2948)
