+++ adl/src/runtime_v2/standing/constants.rs
@@ -1,0 +1,16 @@
+pub const RUNTIME_V2_STANDING_POLICY_SCHEMA: &str = "runtime_v2.standing_policy.v1";
+pub const RUNTIME_V2_STANDING_EVENT_PACKET_SCHEMA: &str = "runtime_v2.standing_event_packet.v1";
+pub const RUNTIME_V2_STANDING_COMMUNICATION_EXAMPLES_SCHEMA: &str =
+    "runtime_v2.standing_communication_examples.v1";
+pub const RUNTIME_V2_STANDING_TRANSITION_PACKET_SCHEMA: &str =
+    "runtime_v2.standing_transition_packet.v1";
+pub const RUNTIME_V2_STANDING_NEGATIVE_CASES_SCHEMA: &str = "runtime_v2.standing_negative_cases.v1";
+
+pub const RUNTIME_V2_STANDING_POLICY_PATH: &str = "runtime_v2/standing/standing_policy.json";
+pub const RUNTIME_V2_STANDING_EVENT_PACKET_PATH: &str = "runtime_v2/standing/standing_events.json";
+pub const RUNTIME_V2_STANDING_COMMUNICATION_EXAMPLES_PATH: &str =
+    "runtime_v2/standing/communication_examples.json";
+pub const RUNTIME_V2_STANDING_TRANSITION_PACKET_PATH: &str =
+    "runtime_v2/standing/standing_transitions.json";
+pub const RUNTIME_V2_STANDING_NEGATIVE_CASES_PATH: &str =
+    "runtime_v2/standing/standing_negative_cases.json";
