+++ adl/src/runtime_v2/standing/mod.rs
@@ -1,0 +1,23 @@
+//! Runtime-v2 standing classes and standing-policy evidence.
+//!
+//! Defines standing classifications, access-rights examples, and negative-case
+//! fixtures for policy- and communication-focused controls.
+
+mod artifacts;
+mod communication;
+mod constants;
+mod events;
+mod negative;
+mod policy;
+mod transition;
+mod validation;
+
+pub use artifacts::RuntimeV2StandingArtifacts;
+pub use communication::{RuntimeV2CommunicationExample, RuntimeV2StandingCommunicationExamples};
+pub use constants::*;
+pub use events::{RuntimeV2StandingEvent, RuntimeV2StandingEventPacket};
+pub use negative::{RuntimeV2StandingNegativeCase, RuntimeV2StandingNegativeCases};
+pub use policy::{RuntimeV2StandingClassPolicy, RuntimeV2StandingPolicy};
+pub use transition::{RuntimeV2StandingTransition, RuntimeV2StandingTransitionPacket};
+
+use super::*;
